compiles on visual studio again

+++ src/DmScanLib.h
@@ -71,10 +71,10 @@ public:
             const unsigned dpi,
             const int brightness,
             const int contrast,
-            const unsigned x1,
-            const unsigned y1,
-            const unsigned x2,
-            const unsigned y2,
+            const float x1,
+            const float y1,
+            const float x2,
+            const float y2,
             const char * filename);
 
     int scanFlatbed(
@@ -87,10 +87,10 @@ public:
             const unsigned dpi,
             const int brightness,
             const int contrast,
-            const unsigned x1,
-            const unsigned y1,
-            const unsigned x2,
-            const unsigned y2,
+            const float x1,
+            const float y1,
+            const float x2,
+            const float y2,
             const DecodeOptions & decodeOptions,
             std::vector<std::unique_ptr<const WellRectangle> > & wellRects);
